revert: one worker file for build
